requirements.txt
django~=3.2.13
psycopg2~=2.9.3
selenium~=3.141.0
requests~=2.27.1
xlrd
PyYAML
lxml
celery==5.1.2
schema~=0.7.4
django-celery-results~=2.2.0
django-filter~=21.1
djangorestframework~=3.13.1
drf-yasg~=1.20.0
Jinja2~=3.0.3
jsonschema~=3.2.0
pydantic~=1.10.1
jsonpath-ng~=1.5.3
rich~=12.5.1
pycrypto==2.6.1
shortuuid~=1.0.9
faker~=13.14.0
aiopg~=1.3.3
aiomysql~=0.1.1
jsf~=0.4.0
redis~=4.3.4
celery~=5.1.2